refactor: predefined commands names
