+++ package.json
@@ -55,7 +55,7 @@
     "katex": "0.16.9",
     "openapi-client-axios": "7.5.4",
     "org-mode-ast": "0.11.2",
-    "orgnote-api": "0.3.0",
+    "orgnote-api": "0.4.1",
     "pinia": "2.1.7",
     "pinia-plugin-persistedstate": "3.2.1",
     "postcss": "8.4.33",
@@ -68,7 +68,7 @@
     "vue": "^3.4.15",
     "vue-i18n": "9.10.1",
     "vue-router": "^4.2.5",
-    "vue3-easy-data-table": "1.5.12",
+    "vue3-easy-data-table": "1.5.47",
     "zod": "3.22.4"
   },
   "devDependencies": {
