--- src/constants/default-commands-meta.ts
@@ -1,15 +1,0 @@
-import { CommandMeta } from 'orgnote-api';
-
-export const TOGGLE_SIDEBAR: CommandMeta = {
-  command: 'toggle sidebar',
-  group: 'global',
-  icon: 'menu',
-  description: 'toggle sidebar',
-};
-
-export const TOGGLE_FILE_MANAGER: CommandMeta = {
-  command: 'toggle file manager',
-  group: 'global',
-  icon: 'folder',
-  description: 'toggle file manager',
-};
