+++ src/constants/index.ts
@@ -1,3 +1,2 @@
 export * from './image-file-extensions';
 export * from './orgnote-schema';
-export * from './default-commands-meta';
