+++ src/hooks/commands/index.ts
@@ -1,0 +1,4 @@
+export * from './route-commands';
+export * from './settings-commands';
+export * from './global-commands';
+export * from './completion-commands';
